+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Ensured all module descriptions in manifest are identical to module descriptions
 
 ## Updated/fixed modules
 - [Manuel Push 2](doc/ManCV.md#manuel-push-2p): In previous release this module was not able to react to trigger-inputs as the manual otherwise stated. There is now a gate/trigger-mode switch next to the input that lets you select whether the input should be detected as a gate or a trigger (defaults to gate as previous version).
+- [Cross-fade switch 4to1](doc/Switch.md#cross-fade-switch-4to1p): Using the context menu you can now specify which volt-level the input should normalize to when not connected (e.g. if you need to switch/fade a fixed volt level to up to 4 different outputs).
+
 
 
 # 2.0.2